% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_ui.R
\name{odin_app}
\alias{odin_app}
\title{Odin Shiny Application}
\usage{
odin_app(model, default_time)
}
\arguments{
\item{model}{An odin model}

\item{default_time}{Default time}
}
\description{
Odin shiny app
}
