% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-model.R
\name{odin_ui_app}
\alias{odin_ui_app}
\title{Interactive odin app}
\usage{
odin_ui_app(model, default_time, title = "odin ui", parameters = NULL,
  extra = NULL, ..., run = TRUE)
}
\arguments{
\item{model}{An odin model (a generator function) created by
\code{odin::odin}.}

\item{default_time}{The default time - this can be either a scalar
number (in which case time is assumed to default to \code{[0,
default_time]} and the starting time is not configurable) or a
length-two vector in which case both the start and end times are
configurable.}

\item{title}{String to use for the application title}

\item{parameters}{A list of parameter information}

\item{extra}{A function to apply over the model output, in order
to produce additional derived columns.  This is designed to
allow more flexible post-run transformations than can easily be
allowed in odin and might be particularly useful in discrete
time stochastic models that are replicated (e.g., to summarise
across model realisations).  Must be provided as a named list of
functions.  The name of the element is the name of the new
variable in the output.}

\item{...}{Additional paramters passed to \code{shiny::runApp}}

\item{run}{Logical, indicating if the app should be run.  If
\code{false} then the app is returned rather than run, and all
arguments in \code{...} are ignored.}
}
\description{
Construct a shiny app from an existing \code{odin} model.
}
