% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-doc.R
\name{odin_ui_doc_app}
\alias{odin_ui_doc_app}
\title{Display model with documentation}
\usage{
odin_ui_doc_app(path_config, ..., run = TRUE)
}
\arguments{
\item{path_config}{Path to a yml file with configuration}

\item{...}{Additional paramters passed to \code{shiny::runApp}}

\item{run}{Logical, indicating if the app should be run.  If
\code{false} then the app is returned rather than run, and all
arguments in \code{...} are ignored.}
}
\description{
Display model with documentation
}
