% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-editor.R
\name{odin_ui_editor_app}
\alias{odin_ui_editor_app}
\title{Shiny app for interactively creating odin models}
\usage{
odin_ui_editor_app(initial_code = NULL, ...)
}
\arguments{
\item{initial_code}{Optional character vector of odin code to be
used as the contents of the editor on startup.  If omitted (or
\code{NULL}) then a very simple default model is used
(exponential growth).}

\item{...}{Additional paramters passed to \code{shiny::runApp}}
}
\description{
Shiny app for interactively creating odin models
}
\details{
This shiny app is subject to all the restrictions that the
\code{\link{odin_ui_app}} shiny app is subject to (no vector
inputs or outputs).
}
