% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-explore.R
\name{odin_ui_explore}
\alias{odin_ui_explore}
\title{Create odin app for exploring a model}
\usage{
odin_ui_explore(path_config, env = .GlobalEnv)
}
\arguments{
\item{path_config}{Path to YAML configuration}

\item{env}{Environment to evaluate any "extra" functions in.  If
these functions are pure (and they should be!) this can be
safely ignored.}
}
\description{
Create an odin app for exploring a model
}
