% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-model.R
\name{odin_ui_app}
\alias{odin_ui_app}
\title{Interactive odin app}
\usage{
odin_ui_app(model, default_time, title = "odin ui", ...)
}
\arguments{
\item{model}{An odin model (a generator function) created by
\code{odin::odin}.}

\item{default_time}{The default time - this can be either a scalar
number (in which case time is assumed to default to \code{[0,
default_time]} and the starting time is not configurable) or a
length-two vector in which case both the start and end times are
configurable.}

\item{title}{String to use for the application title}

\item{...}{Additional arguments passed through to \code{shiny::runApp}}
}
\description{
Construct a shiny app from an existing \code{odin} model.
}
